Handle todo items
