+++ vulkano/src/image/storage.rs
@@ -32,7 +32,7 @@ use crate::memory::pool::PotentialDedicatedAllocation;
 use crate::memory::pool::StdMemoryPool;
 use crate::memory::DedicatedAllocation;
 use crate::memory::ExternalMemoryHandleType;
-use crate::memory::{DeviceMemoryAllocationError, ExternalMemoryHandleTypes};
+use crate::memory::{DeviceMemoryExportError, ExternalMemoryHandleTypes};
 use crate::sync::AccessError;
 use crate::sync::Sharing;
 use crate::DeviceSize;
@@ -227,7 +227,7 @@ impl StorageImage {
 
     /// Exports posix file descriptor for the allocated memory
     /// requires `khr_external_memory_fd` and `khr_external_memory` extensions to be loaded.
-    pub fn export_posix_fd(&self) -> Result<File, DeviceMemoryAllocationError> {
+    pub fn export_posix_fd(&self) -> Result<File, DeviceMemoryExportError> {
         self.memory
             .memory()
             .export_fd(ExternalMemoryHandleType::OpaqueFd)
